+++ changelog.md
@@ -1,5 +1,7 @@
 # Changelog
 
+* 
+
 ## Upcoming
 * Mining charge 
   * Crafting recipe now outputs 2
@@ -8,6 +10,7 @@
 * Potato Seeds > Rooted Potato
 * `survivalreimagined:no_fertility_tooltip` prevents items from displaying the fertility tooltip
 * Experimental music now plays more often 
+* Fixed stamina not following health if the bar was moved (e.g. XP bar was removed)
 
 ## 0.17.1
 * Shields